safley enable swr and make sure to minimalize re renders
