--- web/components/dropdowns/member/index.ts
@@ -1,2 +1,0 @@
-export * from "./project-member";
-export * from "./workspace-member";
