+++ web/components/issues/issue-layouts/roots/cycle-layout-root.tsx
@@ -47,7 +47,8 @@ export const CycleLayoutRoot: React.FC = observer(() => {
           cycleId.toString()
         );
       }
-    }
+    },
+    { revalidateIfStale: false, revalidateOnFocus: false }
   );
 
   const activeLayout = issuesFilter?.issueFilters?.displayFilters?.layout;
