+++ web/components/issues/issue-layouts/roots/module-layout-root.tsx
@@ -43,7 +43,8 @@ export const ModuleLayoutRoot: React.FC = observer(() => {
           moduleId.toString()
         );
       }
-    }
+    },
+    { revalidateIfStale: false, revalidateOnFocus: false }
   );
 
   const userFilters = issuesFilter?.issueFilters?.filters;
