+++ web/components/issues/issue-layouts/roots/project-view-layout-root.tsx
@@ -41,7 +41,8 @@ export const ProjectViewLayoutRoot: React.FC = observer(() => {
           viewId.toString()
         );
       }
-    }
+    },
+    { revalidateIfStale: false, revalidateOnFocus: false }
   );
 
   const issueActions = useMemo(
