+++ web/components/issues/issue-modal/form.tsx
@@ -23,7 +23,7 @@ import {
   ModuleDropdown,
   PriorityDropdown,
   ProjectDropdown,
-  ProjectMemberDropdown,
+  MemberDropdown,
   StateDropdown,
 } from "components/dropdowns";
 // ui
@@ -481,7 +481,7 @@ export const IssueFormRoot: FC<IssueFormProps> = observer((props) => {
                   name="assignee_ids"
                   render={({ field: { value, onChange } }) => (
                     <div className="h-7">
-                      <ProjectMemberDropdown
+                      <MemberDropdown
                         projectId={projectId}
                         value={value}
                         onChange={(assigneeIds) => {
