+++ web/constants/swr-config.ts
@@ -1,7 +1,7 @@
 export const SWR_CONFIG = {
   refreshWhenHidden: false,
-  revalidateIfStale: false,
-  revalidateOnFocus: false,
+  revalidateIfStale: true,
+  revalidateOnFocus: true,
   revalidateOnMount: true,
   refreshInterval: 600000,
   errorRetryCount: 3,
