+++ web/store/module.store.ts
@@ -19,6 +19,7 @@ export interface IModuleStore {
   projectModuleIds: string[] | null;
   // computed actions
   getModuleById: (moduleId: string) => IModule | null;
+  getModuleNameById: (moduleId: string) => string;
   getProjectModuleIds: (projectId: string) => string[] | null;
   // actions
   // fetch
@@ -113,6 +114,13 @@ export class ModulesStore implements IModuleStore {
    * @returns IModule | null
    */
   getModuleById = computedFn((moduleId: string) => this.moduleMap?.[moduleId] || null);
+
+  /**
+   * @description get module by id
+   * @param moduleId
+   * @returns IModule | null
+   */
+  getModuleNameById = computedFn((moduleId: string) => this.moduleMap?.[moduleId]?.name);
 
   /**
    * @description returns list of module ids of the project id passed as argument
